fix: cast numpy array to float32 in read_image() to match training path
When read_image() receives a numpy array (e.g., during inference or
evaluation), the array's native dtype (typically uint8) was passed
through to tf.convert_to_tensor without an explicit cast. This caused
two issues:

1. Dtype mismatch with training path: The file-loading branch already
   produces float32 tensors, so the numpy-input branch was inconsistent,
   leading to potential discrepancies between training and inference.

2. Silent rescaling by tf.image.resize: TensorFlow's resize operation
   auto-scales uint8 images into the [0, 1] range, creating a
   value-range mismatch versus the training pipeline which operates on
   unnormalized float32 values.

Add explicit dtype=tf.float32 to tf.convert_to_tensor to ensure both
code paths produce identically typed and scaled tensors.

Affects inference and evaluation where images are passed as in-memory
numpy arrays. Training is unaffected as it loads from file paths.

+++ base/data/loaders.py
@@ -123,8 +123,9 @@ def read_image(
     """
     try:
         if isinstance(image_input, np.ndarray):
-            # Convert numpy array to tensor
-            image = tf.convert_to_tensor(image_input)
+            # Convert numpy array to tensor (explicit float32 to match training path
+            # and prevent tf.image.resize from auto-scaling uint8 to [0,1])
+            image = tf.convert_to_tensor(image_input, dtype=tf.float32)
             resize_method = 'nearest' if mask else 'bilinear'
             image = tf.image.resize(image, [image_size, image_size], method=resize_method)
         else:
